requirements.txt
install==1.3.5
paho-mqtt==1.5.1
psycopg2==2.9.3
SQLAlchemy==1.4.36
